Feature/kim

+++ src/main/java/com/example/mhbc/controller/BoardController.java
@@ -10,7 +10,7 @@ import com.example.mhbc.repository.*;
 import com.example.mhbc.service.BoardService;
 import com.example.mhbc.service.CommentsService;
 import jakarta.servlet.http.HttpSession;
-import jakarta.transaction.Transactional;
+import org.springframework.transaction.annotation.Transactional;
 import jakarta.validation.Valid;
 import jdk.jshell.execution.Util;
 import lombok.AllArgsConstructor;
@@ -24,6 +24,8 @@ import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -267,14 +269,12 @@ public class BoardController {
 
         int totalCount = (int) paging.getTotalElements();
 
-        Utility.Pagination pagination = new Utility.Pagination(page, itemsPerPage, totalCount, groupSize);
-
-        String link = "/board/oftenquestion_page";
+        Utility.Pagination pagination = new Utility.Pagination(page, itemsPerPage, totalCount, groupSize,"link");
 
         model.addAttribute("paging", paging);
+        model.addAttribute("link","/board/oftenquestion_page");
         model.addAttribute("pagination", pagination);
         model.addAttribute("webtitle", "만화방초 | 자주 묻는 질문");
-        model.addAttribute("link",link);
         model.addAttribute("boardType", boardType);
         model.addAttribute("groupIdx", groupIdx);
 
@@ -326,18 +326,25 @@ public class BoardController {
 
         return "board/personalquestion_page";
     }
+    @Transactional(readOnly = true)
     @RequestMapping("myboard_page")
     public String myboard_page(@RequestParam("board_type") long boardType,
-                                      @RequestParam("group_idx") long groupIdx,
-                                      @RequestParam("member") long memberIdx,
-                                      Model model){
+                                @RequestParam("group_idx") long groupIdx,
+                                  Model model){
 
-        Optional<MemberEntity> member = memberRepository.findById(1L);
-        List<BoardEntity> boardList = boardRepository.findByMemberIdx(1L);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String userid = authentication.getName();
+
+        MemberEntity member = memberRepository.findByUserid(userid);
+
+        List<BoardEntity> boardList = boardRepository.findByMemberIdx(member.getIdx());
+
+        System.out.println("조회된 member idx = " + member.getIdx());
+        System.out.println("조회된 member userid = " + member.getUserid());
 
         model.addAttribute("webtitle", "만화방초 | 내가 작성한 게시글");
         model.addAttribute("boardList", boardList);
-        model.addAttribute("member", member.get());
+        model.addAttribute("member", member);
         model.addAttribute("boardType", boardType);
         model.addAttribute("groupIdx", groupIdx);
 
@@ -400,13 +407,12 @@ public class BoardController {
 
         int totalCount = (int) paging.getTotalElements();
 
-        Utility.Pagination pagination = new Utility.Pagination(page, itemsPerPage, totalCount, groupSize);
+        Utility.Pagination pagination = new Utility.Pagination(page, itemsPerPage, totalCount, groupSize,"link");
 
-        String link = "/board/notice_page";
 
         model.addAttribute("pagination", pagination);
+        model.addAttribute("link","/board/notice_page");
         model.addAttribute("webtitle", "만화방초 | 공지 사항");
-        model.addAttribute("link", link);
         model.addAttribute("paging", paging);
         model.addAttribute("groupIdx", groupIdx);
         model.addAttribute("boardType", boardType);
@@ -497,13 +503,11 @@ public class BoardController {
 
         int totalCount = (int) paging.getTotalElements();
 
-        Utility.Pagination pagination = new Utility.Pagination(page, itemsPerPage, totalCount, groupSize);
-
-        String link = "/board/cmct_page";
+        Utility.Pagination pagination = new Utility.Pagination(page, itemsPerPage, totalCount, groupSize,"link");
 
         model.addAttribute("pagination", pagination);
+        model.addAttribute("link", "/board/cmct_page");
         model.addAttribute("webtitle", "만화방초 | 커뮤니티");
-        model.addAttribute("link", link);
         model.addAttribute("paging", paging);
         model.addAttribute("boardType", boardType);
         model.addAttribute("groupIdx", groupIdx);
@@ -587,7 +591,7 @@ public String comment_proc(@ModelAttribute CommentsDTO commentsDTO,
                            @RequestParam(value = "boardType", required = false) long boardType){
 
 
-    boardService.saveComment(commentsDTO);
+    commentsService.saveComment(commentsDTO,1L);
 
     Long memberIdx = commentsDTO.getMemberIdx();
 
@@ -608,7 +612,7 @@ public String comment_proc(@ModelAttribute CommentsDTO commentsDTO,
         if (attachment.getFilePath() == null || attachment.getFilePath().isEmpty()) {
             throw new RuntimeException("파일 경로가 존재하지 않습니다: " + attachment);
         }
-        String uploadDir = "D:/SpringProject";
+        String uploadDir = "D:/SpringProject/data/";
         Resource resource = new FileSystemResource(uploadDir + attachment.getFilePath());
 
 
@@ -637,7 +641,9 @@ public String comment_proc(@ModelAttribute CommentsDTO commentsDTO,
                          AttachmentEntity attachment,
                          @RequestParam("group_idx") long groupIdx,
                          @RequestParam("board_type") long boardType,
-                         @RequestParam("idx") long boardIdx){
+                         @RequestParam("memberIdx") long memberIdx,
+                         @RequestParam("idx") long boardIdx,
+                         @RequestParam("comments_idx") long commentsIdx){
 
         String redirectUrl = "";
 
@@ -657,11 +663,15 @@ public String comment_proc(@ModelAttribute CommentsDTO commentsDTO,
             redirectUrl = "oftenquestion_page?page=1&board_type="+boardType+"&group_idx="+groupIdx;
         }
 
-        if(boardIdx >= 1) {
-            boardService.deleteBoard(boardIdx);
+        if(boardIdx >= 1 && commentsIdx == 0) {
             utility.deleteAttachments(boardIdx);
+            boardService.deleteBoard(boardIdx);
         }
+        if(commentsIdx >= 1 && boardIdx >= 1){
+            commentsService.deleteComment(commentsIdx);
+            return "redirect:/board/cmct_view?board_type=" + boardType + "&group_idx=" + groupIdx+"&idx="+boardIdx+"&member="+memberIdx;
 
+        }
         return "redirect:/board/"+redirectUrl;
     }
